Applied following changes to branch:
- Switched UPnP command to setting for persistency.
- Switched local "Mono.Nat" to nugget package.
- Moved UPnP functions to its own class.
