+++ Source/Server/Network/Network.cs
@@ -9,26 +9,22 @@ namespace GameServer
 
     public static class Network
     {
-        //AutoPortForwardBool
-        public static bool autoPortForwardSuccessful;
-
         //IP and Port that the connection will be bound to
         private static IPAddress localAddress = IPAddress.Parse(Master.serverConfig.IP);
-        private static int port = int.Parse(Master.serverConfig.Port);
+        public static int port = int.Parse(Master.serverConfig.Port);
 
         //TCP listener that will handle the connection with the clients, and list of currently connected clients
         private static TcpListener connection;
         public static List<ServerClient> connectedClients = new List<ServerClient>();
 
-        private static Punchthrough punchthrough;
-
         //Entry point function of the network class
         public static void ReadyServer()
         {
-            connection = new TcpListener(localAddress, port);
-            connection.Start();
+            if (Master.serverConfig.UseUPnP) { _ = new UPnP(); }
 
             Threader.GenerateServerThread(Threader.ServerMode.Sites);
+            connection = new TcpListener(localAddress, port);
+            connection.Start();
 
             Logger.WriteToConsole("Type 'help' to get a list of available commands", Logger.LogMode.Warning);
             Logger.WriteToConsole($"Listening for users at {localAddress}:{port}", Logger.LogMode.Warning);
@@ -93,57 +89,5 @@ namespace GameServer
                 Logger.WriteToConsole($"Error disconnecting user {client.username}, this will cause memory overhead", Logger.LogMode.Warning);
             }
         }
-
-        public static void TryToForwardPort()
-        {
-            autoPortForwardSuccessful = false;
-            Logger.WriteToConsole($"Attempting to forward port {port}");
-            Task.Run(delegate { punchthrough = new Punchthrough(); });
-
-            //Wait for the server to try to portforward for 5 seconds or until it is successful
-            Thread.Sleep(5000);
-            if (!autoPortForwardSuccessful)
-            {
-                Logger.WriteToConsole("Could not Auto PortForward - \n " +
-                    "              Possible causes:\n" +
-                    "              - the port is being used\n" +
-                    "              - the router has uPnP disabled\n" +
-                    "              - or the router/modem does not have ports", Logger.LogMode.Warning);
-                NatUtility.StopDiscovery();
-            }
-        }
-
-        //class for auto portforwarding with UPnP
-        public class Punchthrough
-        {
-
-            public Punchthrough()
-            {
-                Logger.WriteToConsole("Attempting to auto portforward");
-                NatUtility.DeviceFound += DeviceFound;
-                NatUtility.DeviceLost += DeviceLost;
-                NatUtility.StartDiscovery();
-            }
-
-            private void DeviceFound(object sender, DeviceEventArgs args)
-            {
-                Logger.WriteToConsole("Device found");
-                autoPortForwardSuccessful = true;
-                INatDevice device = args.Device;
-                device.CreatePortMap(new Mapping(Protocol.Tcp, port, port));
-                Logger.WriteToConsole("portforward successful");
-                // on device found code
-            }
-
-            private void DeviceLost(object sender, DeviceEventArgs args)
-            {
-                Logger.WriteToConsole("Device Lost");
-                INatDevice device = args.Device;
-                device.DeletePortMap(new Mapping(Protocol.Tcp, port, port));
-                // on device disconnect code
-            }
-        }
     }
-
-    
 }
